+++ spinta/core/ufuncs.py
@@ -285,6 +285,9 @@ class Bind(Unresolved):
     def __repr__(self):
         return self.name
 
+    def __str__(self):
+        return self.name
+
 
 class Pair(Unresolved):
 
@@ -308,6 +311,9 @@ class Positive(Bind):
 class GetAttr(Unresolved):
     obj: str
     name: Union[GetAttr, Bind]
+
+    def __str__(self):
+        return f'{self.obj}.{str(self.name)}'
 
 
 bind = functools.partial(Expr, 'bind')