+++ spinta/utils/data.py
@@ -1,9 +1,10 @@
 from collections.abc import Iterable
 
+from spinta.utils.nestedstruct import split_excluding_parentheses
 from spinta.utils.schema import NA
 
 
-def take(keys, *args):
+def take(keys, *args, separator='.'):
     """Ultimate data extraction function
 
     Examples:
@@ -67,7 +68,7 @@ def take(keys, *args):
             if v is not NA and (reserved or not k.startswith('_'))
         }
     else:
-        keys = [k.split('.') for k in keys]
+        keys = [split_excluding_parentheses(k, separator=separator) for k in keys]
 
     data = {}
     for k in keys:
@@ -81,7 +82,7 @@ def take(keys, *args):
                 if v is not NA:
                     if key:
                         return v
-                    data['.'.join(k)] = v
+                    data[separator.join(k)] = v
                     break
 
     if key: